+++ CHANGELOG.md
@@ -5,7 +5,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 This project **DOES NOT** adhere to [Semantic Versioning](http://semver.org/spec/v2.0.0.html) yet.
 
 ## [Unreleased]
-[Unreleased]: https://bitbucket.org/atlassian/db-replica/branches/compare/master%0Drelease-0.1.8
+[Unreleased]: https://bitbucket.org/atlassian/db-replica/branches/compare/master%0Drelease-0.1.9
+
+## [0.1.9] - 2020-11-25
+[0.1.9]: https://bitbucket.org/atlassian/db-replica/branches/compare/release-0.1.9%0Drelease-0.1.8
 
 ### Fixed
 - Assign deletes to the main connection for `executeQuery` calls